Fixing incorrect bit placement for PLL multiplier in variant.cpp
The PLL multiplier  in the STM chip only goes up to 16. GD32 based chips can go higher.  To do so they define specific register bit for the 4th bit (and a 5th for GD32F30x).
For GD32F10x the 4th bit is placed in reg bit 27.

This change fixes the clock register, which was likely operating at a slow clock due to this bug.